analy_app/requirements.txt
absl-py @ file:///C:/b/abs_14wchyjaia/croot/absl-py_1714140489450/work
altair==4.2.2
altgraph @ file:///C:/b/abs_f2edualeyv/croot/altgraph_1670426107695/work
annotated-types==0.7.0
anywidget==0.9.13
appdirs==1.4.4
APScheduler==3.11.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
attrs==24.3.0
backoff==2.2.1
blinker==1.9.0
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.5.0
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
comm==0.2.2
dateutils==0.6.12
decorator==5.1.1
duckdb==1.1.3
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio @ file:///C:/b/abs_7d9olmtnd2/croot/grpc-split_1716834609196/work
gw_dsl_parser==0.1.49.1
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///C:/b/abs_9d0lp1ysmt/croot/importlib_metadata-suite_1732633510972/work
importlib_resources==6.5.2
ipylab==1.0.0
ipython==8.18.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2 @ file:///C:/b/abs_ddyeze2h6_/croot/jinja2_1730903330414/work
joblib @ file:///C:/b/abs_f4b98l6lgk/croot/joblib_1718217224240/work
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kanaries_track==0.0.5
Markdown @ file:///C:/b/abs_98lv_ucina/croot/markdown_1671541919225/work
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
monotonic==1.6
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
narwhals==1.21.1
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==1.26.4
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.2.3
parso==0.8.4
pefile @ file:///C:/b/abs_68oy9ka6fm/croot/pefile_1720651782540/work
pillow @ file:///C:/b/abs_2cm0qc1vxk/croot/pillow_1731594717982/work
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.1.1
psygnal==0.11.1
pure_eval==0.2.3
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pygwalker==0.4.9.13
pyinstaller @ file:///C:/b/abs_09h007ff52/croot/pyinstaller_1721682848440/work
pyinstaller-hooks-contrib @ file:///C:/b/abs_deumjuw0e0/croot/pyinstaller-hooks-contrib_1721675898298/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywin32-ctypes @ file:///C:/b/abs_2cfx5l4nvi/croot/pywin32-ctypes_1709340246092/work
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
quickjs==1.19.4
referencing==0.35.1
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich==13.9.4
rpds-py==0.22.3
scikit-learn @ file:///C:/b/abs_69a_gdu9bf/croot/scikit-learn_1734547445211/work/dist/scikit_learn-1.5.2-cp39-cp39-win_amd64.whl#sha256=8ca654624673ea6d085304ada0d5bfdf8ca5cbf947f00011e9c49489025899b2
scipy @ file:///C:/Users/dev-admin/cr_scipy-feedstock/scipy_1733774324044/work/dist/scipy-1.13.1-cp39-cp39-win_amd64.whl#sha256=f7218697f964ae4372cdcb6f20a05db0d338ccc574c96e9dbf235ed6547dd462
segment-analytics-python==2.2.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
SQLAlchemy==2.0.36
sqlglot==26.1.2
st-pages==1.0.1
stack-data==0.6.3
streamlit==1.41.1
streamlit-aggrid==1.0.5
streamlit-elements==0.1.0
sympy @ file:///C:/b/abs_4e4p71hdj_/croot/sympy_1724938208509/work
tenacity==9.0.0
tensorboard @ file:///C:/b/abs_8bayrkofnj/croot/tensorboard_1723042051994/work/tensorboard-2.17.0-py3-none-any.whl#sha256=859a499a9b1fb68a058858964486627100b71fcb21646861c61d31846a6478fb
tensorboard_data_server @ file:///C:/b/abs_96oxnengxy/croot/tensorboard-data-server_1724172563443/work/tensorboard_data_server-0.7.0-py3-none-any.whl#sha256=e1491598092b1fe59a6545fb3d1079464cf3a999105f6bdd8a2ec8af25cf8d4c
threadpoolctl @ file:///C:/b/abs_def0dwqlft/croot/threadpoolctl_1719407816649/work
toml==0.10.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
wasmtime==28.0.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug @ file:///C:/b/abs_b21bhstb1q/croot/werkzeug_1731005800632/work
widgetsnbextension==4.0.13
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
zipp @ file:///C:/b/abs_2fcpcg89my/croot/zipp_1732630746332/work
